+++ ontimize/components/table/column/cell-renderer/percentage/o-table-cell-renderer-percentage.component.ts
@@ -1,25 +1,28 @@
 import { Component, Injector, OnInit, TemplateRef, ViewChild } from '@angular/core';
 
 import { NumberService } from '../../../../../services';
-import { IPercentPipeArgument, OPercentPipe } from '../../../../../pipes';
+import { IPercentPipeArgument, OPercentPipe, OPercentageValueBaseType } from '../../../../../pipes';
 import { DEFAULT_INPUTS_O_TABLE_CELL_RENDERER_REAL, OTableCellRendererRealComponent } from '../real/o-table-cell-renderer-real.component';
 
 export const DEFAULT_INPUTS_O_TABLE_CELL_RENDERER_PERCENTAGE = [
-  ...DEFAULT_INPUTS_O_TABLE_CELL_RENDERER_REAL
+  ...DEFAULT_INPUTS_O_TABLE_CELL_RENDERER_REAL,
+  'valueBase: value-base'
 ];
 
 @Component({
   moduleId: module.id,
   selector: 'o-table-cell-renderer-percentage',
   templateUrl: './o-table-cell-renderer-percentage.component.html',
-  inputs: DEFAULT_INPUTS_O_TABLE_CELL_RENDERER_REAL
+  inputs: DEFAULT_INPUTS_O_TABLE_CELL_RENDERER_PERCENTAGE
 })
 export class OTableCellRendererPercentageComponent extends OTableCellRendererRealComponent implements OnInit {
 
-  public static DEFAULT_INPUTS_O_TABLE_CELL_RENDERER_REAL = DEFAULT_INPUTS_O_TABLE_CELL_RENDERER_REAL;
+  public static DEFAULT_INPUTS_O_TABLE_CELL_RENDERER_PERCENTAGE = DEFAULT_INPUTS_O_TABLE_CELL_RENDERER_PERCENTAGE;
 
   protected decimalSeparator: string = '.';
   protected decimalDigits: number = 0;
+  protected valueBase: OPercentageValueBaseType = 1;
+
   protected numberService: NumberService;
 
   protected componentPipe: OPercentPipe;
@@ -48,7 +51,8 @@ export class OTableCellRendererPercentageComponent extends OTableCellRendererRea
       maxDecimalDigits: this.decimalDigits,
       decimalSeparator: this.decimalSeparator,
       grouping: this.grouping,
-      thousandSeparator: this.thousandSeparator
+      thousandSeparator: this.thousandSeparator,
+      valueBase: this.valueBase
     };
   }